+++ src/term.h
@@ -34,6 +34,7 @@ typedef enum {
 
 int terminal_mode(PROGRAMMER * pgm, struct avrpart * p);
 char * terminal_get_input(const char *prompt);
+void terminal_setup_update_progress();
 
 #ifdef __cplusplus
 }
